ecomm_app/urls.py
from django.urls import path, include
from ecomm_app.views import  UserViewSet, ProductViewSet
from rest_framework import routers

router = routers.DefaultRouter()
router.register(r'users', UserViewSet, basename='user')
router.register(r'products', ProductViewSet, basename='product')

urlpatterns = [
    path('', include(router.urls)),
]