requirements.txt
Django==4.2.6
djangorestframework==3.14.0
mysqlclient==2.0.1